M1: widen CRAM positions and contig lengths to u64
Eliminate the silent ``as u32`` truncation flagged in the code review:
``MappedRead.pos``, ``MappedRead.mate_pos``, ``ContigEntry.length``,
``WindowKey.pos``, ``HeadKey.pos``, ``PerFileOrder``, ``window_anchor``,
``DuplicateReadAcrossFiles.pos``, and the matching test fixtures
(``RecordSpec``, ``ContigSpec``, ``HeaderOverrides``) are now ``u64``.
Adds a regression test that round-trips a position above ``u32::MAX``.

+++ src/per_sample_caller/errors.rs
@@ -75,7 +75,7 @@ pub enum CramInputError {
         path_a: PathBuf,
         path_b: PathBuf,
         ref_id: usize,
-        pos: u32,
+        pos: u64,
     },
 
     #[error("malformed record in '{path}' (qname='{qname}'): {source}")]
